rosatom-backend/requirements.txt
alembic==1.4.3
aniso8601==8.0.0
bcrypt==3.2.0
cffi==1.14.4
cheroot==8.4.8
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.text==3.2.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
more-itertools==8.6.0
portend==2.7.0
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
pywin32==300
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.36.8
tempora==4.0.1
Werkzeug==1.0.1
zc.lockfile==2.0
